[refacotr]rename views => pages
